Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int x,y,z,avg;
+    double x,y,z,avg;
     scanf("%lf %lf %lf",&x, &y, &z);
-    avg = x+y+x/3;
-    printf("Average: %.2lf",avg);
+    avg = (x+y+x)/3;
+    printf("Average: %.2f",avg);
     return 0;
 }
